renamed XcodeRunner as XcodeBuildRunner added XcodeRunner for xcrun added Notarize command made Compress command compress into .build/compressed by default made --to and --latest arguments optional for compress command

+++ Sources/ReleaseTools/CompressCommand.swift
@@ -9,14 +9,17 @@ import Runner
 
 extension Result {
     static let infoUnreadable = Result(400, "Couldn't read archive info.plist.")
+    static let compressFailed = Result(401, "Compression failed.")
 }
 
 class CompressCommand: Command {
+    static let compressedPath = ".build/compressed"
+
     override var description: Command.Description {
         return Description(
             name: "compress",
             help: "Compress the output of the export command for distribution.",
-            usage: ["--to=<to> --latest=<latest>"],
+            usage: ["[[--to=<to>] --latest=<latest>]"],
             options: ["--repo=<repo>": "The repository containing the appcast and updates."],
             returns: [.infoUnreadable]
         )
@@ -29,20 +32,21 @@ class CompressCommand: Command {
 
         let exportedAppPath = URL(fileURLWithPath: ExportCommand.exportPath).appendingPathComponent(archive.name)
         let ditto = Runner(for: URL(fileURLWithPath: "/usr/bin/ditto"))
-        let archiveFolder = try shell.arguments.expectedOption("to")
+        let archiveFolder = shell.arguments.option("to", default: CompressCommand.compressedPath)
         let destination = URL(fileURLWithPath: archiveFolder).appendingPathComponent(archive.versionedZipName)
         
         shell.log("Compressing \(archive.name) to \(archiveFolder) as \(archive.versionedZipName).")
         let result = try ditto.sync(arguments: ["-c", "-k", "--sequesterRsrc", "--keepParent", exportedAppPath.path, destination.path])
         if result.status != 0 {
-            return Result.exportFailed.adding(supplementary: result.stderr)
+            return Result.compressFailed.adding(supplementary: result.stderr)
         }
         
-        let latestFolder = try shell.arguments.expectedOption("latest")
-        shell.log("Saving copy of archive to \(latestFolder) as \(archive.unversionedZipName).")
-        let latestZip = URL(fileURLWithPath: latestFolder).appendingPathComponent(archive.unversionedZipName)
-        try? FileManager.default.removeItem(at: latestZip)
-        try FileManager.default.copyItem(at: destination, to: latestZip)
+        if let latestFolder = shell.arguments.option("latest") {
+            shell.log("Saving copy of archive to \(latestFolder) as \(archive.unversionedZipName).")
+            let latestZip = URL(fileURLWithPath: latestFolder).appendingPathComponent(archive.unversionedZipName)
+            try? FileManager.default.removeItem(at: latestZip)
+            try FileManager.default.copyItem(at: destination, to: latestZip)
+        }
         
         return .ok
     }
